providers: update IUserConnector.SetLeverage interface

+++ finance/Providers/src/Annium.Finance.Providers.Abstractions.Connectors/Connectors/IUserConnector.cs
@@ -11,7 +11,7 @@ public interface IUserConnector : IConnectorBase
     ITableView<AssetDto> Assets { get; }
     ITableView<PositionDto> Positions { get; }
     ITableView<OrderDto> Orders { get; }
-    Task<UserResult> SetLeverage(PositionDto position, byte leverage);
+    Task<UserResult> SetLeverage(PositionDto position, decimal leverage);
     Task<UserResult<OrderDto>> InitOrder(IInitOrderRequest request);
     Task<UserResult<OrderDto>> ModifyOrder(IModifyOrderRequest request);
     Task<UserResult> CancelOrder(OrderDto order);
